examples/qt05-examples-benchqc/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qt05-examples-benchqc"
version = "0.1.0"
description = "example using benchqc to compare quantum and classical molecular calculations"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "lwfm@git+https://github.com/lwfm-proj/lwfm@develop",
    "numpy==1.23.5",
    "jarvis-tools",
    "qiskit==0.43.1",
    "qiskit_nature==0.6.2",
    "pyscf",
    "qiskit-aer==0.12.0",
    ]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
